% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_signature}
\alias{bootstrap_signature}
\title{Perform Bootstrapping for Signature Creation}
\usage{
bootstrap_signature(
  data,
  features,
  exposure,
  n_boot,
  covars = c(),
  folds = 5,
  parallel = F,
  cores = 18,
  ...
)
}
\arguments{
\item{data}{A data frame or matrix containing the dataset. Rows represent samples, and columns represent features, covariates, and the exposure variable.}

\item{features}{A character vector of column names in \code{data} representing the features to include in the model.}

\item{exposure}{A character string representing the column name in \code{data} corresponding to the exposure variable.}

\item{n_boot}{An integer specifying the number of bootstrap iterations.}

\item{covars}{A character vector of column names in \code{data} representing additional covariates to include in the model. Defaults to an empty vector.}

\item{folds}{An integer specifying the number of folds for cross-validation. Defaults to 5.}

\item{parallel}{A logical value indicating whether to use parallel processing. Defaults to \code{FALSE}.}

\item{cores}{An integer specifying the number of CPU cores to use if \code{parallel = TRUE}. Defaults to 18.}
}
\value{
A list with the following components:
\describe{
\item{bootstrap_coefs_l1se}{A sparse matrix containing bootstrap counts for each feature selected using the 1-SE rule.}
\item{bootstrap_coefs_lmin}{A sparse matrix containing bootstrap counts for each feature selected using the minimum lambda.}
\item{bootstrap_cors}{A data frame containing the number of selected features and the correlations between predicted signatures and the exposure variable for each bootstrap iteration.}
}
}
\description{
This function performs bootstrapping to evaluate the stability of feature selection and signature performance
in a cross-validation setting using LASSO or a similar approach.
}
\examples{
\dontrun{
# Example dataset
data <- data.frame(matrix(rnorm(1000), nrow = 100, ncol = 10))
colnames(data) <- c(paste0("feature", 1:8), "exposure", "covar")
results <- bootstrap_signature(
  data = data,
  features = paste0("feature", 1:8),
  exposure = "exposure",
  n_boot = 100,
  covars = "covar",
  folds = 5,
  parallel = TRUE,
  cores = 4
)
}

}
