% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covar_to_numeric}
\alias{covar_to_numeric}
\title{Convert Covariates to Numeric}
\usage{
covar_to_numeric(data, covars)
}
\arguments{
\item{data}{A data frame containing the columns to be converted.}

\item{covars}{A character vector or numeric vector specifying the columns to convert to numeric.}
}
\value{
A data frame with the specified columns converted to numeric.
}
\description{
This function converts specified columns in a data frame to numeric.
}
\examples{
# Example with character vector of column names
data <- data.frame(a = c("1", "2", "3"), b = c("4", "5", "6"), c = c(TRUE, FALSE, TRUE))
covar_to_numeric(data, c("a", "b"))

# Example with numeric indices
data <- data.frame(a = c("1", "2", "3"), b = c("4", "5", "6"), c = c(TRUE, FALSE, TRUE))
covar_to_numeric(data, c(1, 2))

}
