% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signature.R
\name{create_signature}
\alias{create_signature}
\title{Create Omics-Based Exposure Signature Using Cross-Validated LASSO}
\usage{
create_signature(
  data,
  train_idx,
  features,
  exposure,
  covars = c(),
  folds = 5,
  parallel = F,
  cores = 18
)
}
\arguments{
\item{data}{A data frame containing both features and the outcome variable.}

\item{train_idx}{A vector of row indices specifying the training subset of \code{data}.}

\item{features}{A character vector of column names specifying the features (predictor variables) to be included in the model.}

\item{folds}{Integer specifying the number of folds for cross-validation. Default is 5.}

\item{parallel}{Logical value indicating whether to use parallel processing. Default is \code{FALSE}.}

\item{cores}{Integer specifying the number of cores to use if \code{parallel = TRUE}. Default is 18.}

\item{outcome}{A character string specifying the name of the outcome variable (response) in \code{data}.}
}
\value{
A list with two elements:
\itemize{
\item \code{crossval}: The fitted \code{cv.glmnet} object.
\item \code{results}: A list of selected variables based on \code{lambda.min} and \code{lambda.1se}.
}
}
\description{
Fits a cross-validated LASSO model using \code{cv.glmnet} to identify a signature of features associated with a specified outcome. Optionally supports parallel processing for faster computation.
}
\details{
This function performs LASSO regression using cross-validation to select optimal features. If \code{parallel = TRUE}, the function uses \code{doMC} for parallel computation with the specified number of cores.
}
\examples{
\dontrun{
data <- my_data_frame
train_index <- sample(1:nrow(data), size = floor(0.8 * nrow(data)))
feature_names <- colnames(data)[2:100] # Example feature names
outcome_name <- "my_outcome"
signature <- create_signature(data, train_index, feature_names, outcome_name, n_folds = 10, parallel = TRUE, n_cores = 4)
}

}
\seealso{
\code{\link[glmnet]{cv.glmnet}}, \code{\link[doMC]{registerDoMC}}, \code{\link{format_cvglmnet}}
}
