% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signature.R
\name{format_cvglmnet}
\alias{format_cvglmnet}
\title{Format cv.glmnet Object from create_signature}
\usage{
format_cvglmnet(df_crossval, features)
}
\arguments{
\item{df_crossval}{A \code{cv.glmnet} object, typically obtained from \code{create_signature}.}
}
\value{
A list with two elements:
\itemize{
\item \code{crossval}: The original \code{cv.glmnet} object.
\item \code{results}: A list with selected variables for \code{lambda.min} and \code{lambda.1se}.
}
}
\description{
Extracts selected features from a \code{cv.glmnet} object using the lambda values for minimum error (\code{lambda.min}) and one standard error (\code{lambda.1se}).
}
