% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Calculate Residuals for Features Adjusted for Covariates and Exposure}
\usage{
get_residuals(
  data,
  features,
  covars,
  exposure,
  id_cols,
  ctrl_idx = c(),
  parallel = F,
  cores = 18
)
}
\arguments{
\item{data}{A data frame containing the features, covariates, and exposure variables.}

\item{features}{A character vector of feature names (e.g., metabolite names) or a numeric vector of column indices for which residuals are to be calculated.}

\item{covars}{A character vector of covariate column names or a numeric vector of column indices to adjust for in the linear models.}

\item{exposure}{A character string specifying the exposure variable column name or a numeric index of the exposure variable to include in the models.}

\item{id_cols}{A character vector of column names or a numeric vector of column indices to use as identifiers (e.g., sample IDs) in the resulting data frame.}

\item{ctrl_idx}{A numeric vector specifying the row indices of control samples. If empty, all rows are treated as controls.}

\item{parallel}{A logical value indicating whether to use parallel processing. Defaults to \code{FALSE}.}

\item{cores}{An integer specifying the number of cores to use for parallel processing. Defaults to 18.}
}
\value{
A data frame containing the residuals for each feature, along with the identifier columns and remaining data columns.
}
\description{
This function computes residuals for a given set of features after adjusting
for specified covariates and an exposure variable. It uses linear models for adjustment
and supports parallel processing to improve performance when analyzing multiple features.
}
\details{
The function fits a linear model for each feature using the formula:
\code{`feature_name ~ exposure + covariates`}.
Residuals are computed for control samples (if specified) and optionally for case samples.
Residuals are adjusted by adding back the fixed effect of the exposure variable.

If \code{ctrl_idx} is provided, the linear model is fit only on the control samples, and residuals computed for the case samples using the control model.
}
\examples{
# Example dataset
data <- data.frame(
  ID = 1:100,
  feature1 = rnorm(100),
  feature2 = rnorm(100),
  covar1 = runif(100),
  covar2 = runif(100),
  exposure = sample(c(0, 1), 100, replace = TRUE)
)

# Specify parameters
features <- c("feature1", "feature2")
covars <- c("covar1", "covar2")
exposure <- "exposure"
id_cols <- "ID"

# Compute residuals
residuals <- get_residuals(
  data = data,
  features = features,
  covars = covars,
  exposure = exposure,
  id_cols = id_cols,
  parallel = FALSE
)

}
