% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_utils.R
\name{pearson_cor}
\alias{pearson_cor}
\title{Compute Pearson correlations}
\usage{
pearson_cor(data, feat, exposure, covars)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{feat}{A string specifying the feature of interest.}

\item{exposure}{A string specifying the exposure variable.}

\item{covars}{A character vector of covariate names to adjust for.}
}
\value{
A named vector with:
\itemize{
\item \code{feat}: the feature name,
\item \code{coef}: the estimated coefficient for the exposure variable,
\item \code{pval}: the p-value associated with the exposure variable,
\item \code{type}: the string "pearson".
}
}
\description{
This function calculates the Pearson correlation between a feature and an exposure, adjusting for covariates.
}
\details{
This function fits a linear model of the form: \code{feat ~ exposure + covariates}.
The estimated coefficient and p-value for the exposure variable are returned as a measure of
correlation, adjusted for covariates.
}
\examples{
pearson_cor(data = my_data, feat = "metabolite1", exposure = "alcohol", covars = c("age", "sex"))

}
