% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_signature.R
\name{pred_signature}
\alias{pred_signature}
\title{Predict Exposure Signatures from Cross-Validated Model}
\usage{
pred_signature(data, crossval, test_idx, features, exposure)
}
\arguments{
\item{data}{A data frame or matrix containing the full set of features and exposure variable.}

\item{crossval}{A \code{cv.glmnet} object resulting from cross-validation that contains the fitted model.}

\item{test_idx}{A vector of indices specifying the set for which predictions are to be made.}

\item{features}{A character vector or column indices specifying the feature names used for prediction.}

\item{exposure}{A character string specifying the name of the exposure (response variable) in the \code{data} used for comparison with predictions.}
}
\value{
A list containing two matrices:
\itemize{
\item \code{l1se}: A matrix with the actual exposure data and predicted signature using \code{lambda.1se}.
\item \code{lmin}: A matrix with the actual exposure data and predicted signature using \code{lambda.min}.
}
Each matrix has the exposure values in the first column and the predicted signature values in the second column.
}
\description{
This function uses a cross-validated model object (\code{cv.glmnet}) to predict exposure signatures for a test set. It returns the predicted values for both the \code{lambda.min} and \code{lambda.1se} values from the cross-validation.
}
\details{
The function predicts exposure signatures using the fitted \code{cv.glmnet} model. Predictions are made for both the \code{lambda.1se} and \code{lambda.min} values. The returned matrices combine actual exposure values and predicted signatures for comparison.
}
\examples{
\dontrun{
# Example usage
data <- your_data_frame
crossval <- cv.glmnet(x = data[train_idx, features], y = data[train_idx, exposure])
test_idx <- sample(1:nrow(data), 50)  # Specify test set indices
pred_results <- pred_signature(data, crossval, test_idx, features, exposure)
}

}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
